hungerDelight/requirements.txt
Django==3.1.1
djangorestframework==3.12.1
mysqlclient==2.0.1
